pyproject.toml
[tool.poetry]
name = "student-api-server"
version = "0.1.0"
description = ""
authors = ["Gudla Devi Shreya <23f3003943@ds.study.iitm.ac.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.119.0"
uvicorn = "^0.37.0"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
pydantic = "^2.12.2"
celery = "^5.5.3"
redis = "^6.4.0"
python-dotenv = "^1.1.1"
requests = "^2.32.5"
PyGithub = "^2.8.1"
tenacity = "^9.1.2"
email-validator = "^2.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"